% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dot}
\alias{dot}
\title{Print a dot every k loops}
\usage{
dot(k, every = 100, symbol = ".")
}
\arguments{
\item{k}{looping paramters}

\item{every}{print a symbol when k mod every == 0 (default 100)}

\item{symbol}{String to print (default \code(.))}
}
\description{
Print a dot every k loops
}
\examples{
for( k in 1:1000 ){
  dot(k)
  dot(k, every=150, symbol="_")
}
}

