% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCNArmSummary}
\alias{getCNArmSummary}
\title{Make a table of CN summary stats per chromosome arm}
\usage{
getCNArmSummary(segments, bandsFile)
}
\arguments{
\item{segments}{Data table with segments from qdnaseq.}

\item{arms}{Data table with chromosome arms.}
}
\value{
A summary table per chromosome arm
}
\description{
Make a table of CN summary stats per chromosome arm
}
\examples{
#getCNArmSummary(segments, bandsFile = "~/genome/cytobands.txt")
}

