% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCombinedMaf}
\alias{getCombinedMaf}
\title{Get a combined MAF from a set of reports}
\usage{
getCombinedMaf(reports)
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A combined MAF data table
}
\description{
Get a combined MAF from a set of reports
}
\examples{
# getCombinedMaf(reports)
}

