% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGeneBodyCoverage}
\alias{getGeneBodyCoverage}
\title{Get normalized gene body coverage for reports}
\usage{
getGeneBodyCoverage(reports, src = "StarRNASeqMetrics")
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A data frame with normalized gene body coverage for all samples
}
\description{
Get normalized gene body coverage for reports
}
\examples{
# dat <- getGeneBodyCoverage(reports)
# ggplot(dat, aes(x=NORM_POS, y=NORM_COV, colour=DataReportID)) + geom_line()
}

