% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGeneExprMatrix.R
\name{getGeneExprMatrix}
\alias{getGeneExprMatrix}
\title{Get gene expression from htseq-count}
\usage{
getGeneExprMatrix(reports)
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A data frame with counts from htseq-count
}
\description{
Get gene expression from htseq-count
}
\examples{
#dat <- getGeneExprMatrix(reports)
}

