% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHSMetrics}
\alias{getHSMetrics}
\title{Get HS Metrics}
\usage{
getHSMetrics(reports, src = "RNASeqCapHsMetrics")
}
\arguments{
\item{reports}{data frame with reports}

\item{src}{string with name of column name to use (from reports)}
}
\value{
A data table with hs metrics
}
\description{
Get HS Metrics
}
\examples{
#dat <- getHSMetrics(reports, src="RNASeqCapHsMetrics")
#dat <- getHSMetrics(reports, src="TumorHsMetrics")
#dat <- getHSMetrics(reports, src="NormalHsMetrics")
}

