% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getIsoformExprMatrix.R
\name{getIsoformExprMatrix}
\alias{getIsoformExprMatrix}
\title{Get isoform expression from sailfish output}
\usage{
getIsoformExprMatrix(reports, columnToUse = "EstimatedNumKmers")
}
\arguments{
\item{reports}{data frame with reports}

\item{columnToUse}{Name of the column in sailfish output to use}
}
\value{
A data frame with isoform quant from sailfish base on columnToUse
}
\description{
Get isoform expression from sailfish output
}
\examples{
#dat <- getIsoformExprMatrix(reports)
#dat <- getIsoformExprMatrix(reports, columnToUse="TPM")
}

