% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMarkDupsMetrics}
\alias{getMarkDupsMetrics}
\title{Get MarkDups Metrics}
\usage{
getMarkDupsMetrics(reports, src = "PANEL_MARKDUPS_METRICS")
}
\arguments{
\item{reports}{data frame with reports}

\item{src}{string with name of column name to use (from reports)}
}
\value{
A data table with metrics from markDuplicates
}
\description{
Get MarkDups Metrics
}
\examples{
#dat <- getMarkDupsMetrics(reports, src="PANEL_MARKDUPS_METRICS")
#dat <- getMarkDupsMetrics(reports, src="WGS_MARKDUPS_METRICS")
}

