% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOncofuse}
\alias{getOncofuse}
\title{Get Oncofuse results}
\usage{
getOncofuse(reports)
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A data table with Oncofuse results
}
\description{
Get Oncofuse results
}
\examples{
#dat <- getOncofuse(reports, src="RNASeqCapHsMetrics")
}

