% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getReadcountPerChr.R
\name{getReadcountPerChr}
\alias{getReadcountPerChr}
\title{Get read count per chromsome}
\usage{
getReadcountPerChr(reports)
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A data frame with read counts per chr
}
\description{
Get read count per chromsome
}
\examples{
#dat <- getReadcountPerChr(reports)
}

