% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getReports}
\alias{getReports}
\title{Get a data frame with all reports found in path}
\usage{
getReports(path, recursive = TRUE)
}
\arguments{
\item{path}{Path to directory with reports.}

\item{recursive}{Boolean whether or not to recurse when looking for report files (default TRUE)}
}
\value{
A data frame with reports
}
\description{
Get a data frame with all reports found in path
}
\examples{
getReports("/home/Crisp/clinseq/BREASTv4")
}

