% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRNAseqMetrics.R
\name{getRNAseqMetrics}
\alias{getRNAseqMetrics}
\title{Get normalized gene body coverage for reports}
\usage{
getRNAseqMetrics(reports, src = "RNASEQ_RNASEQ_METRICS")
}
\arguments{
\item{reports}{data frame with reports}

\item{src}{string with name of column name to use (from reports)}
}
\value{
A data frame with normalized gene body coverage for all samples
}
\description{
Get normalized gene body coverage for reports
}
\examples{
#dat <- getRNAseqMetrics(reports)
#ggplot(subset(dat, METRIC=="PCT_RIBOSOMAL_BASES"),
#       aes(x=DataReportID, y=VALUE, fill=DataReportID)) +
#geom_bar(stat="identity")
}

