% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeEmptyDataTable}
\alias{makeEmptyDataTable}
\title{Make an empty data.table from a vector of column names}
\usage{
makeEmptyDataTable(header)
}
\arguments{
\item{header}{Vector with column names}
}
\description{
Make an empty data.table from a vector of column names
}
\examples{
makeEmptyDataTable(c("Column1", "Column2"))
}

