% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL,
  panelnames = NULL, panelnames.size = 16)
}
\arguments{
\item{...}{plots to draw}

\item{plotlist}{list of plots to draw}

\item{cols}{number of columns}

\item{layout}{matrix with layout specifications}

\item{panelnames}{array of panel names (default LETTERS)}

\item{panelnames.size}{size of panel names (default 16)}
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
- cols:   Number of columns in layout
- layout: A matrix specifying the layout. If present, 'cols' is ignored.
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\examples{
# getCombinedMaf(reports)
library(ggplot2)
p1 <- ggplot(mtcars, aes(factor(cyl), mpg)) + geom_boxplot()
p2 <- ggplot(mtcars, aes(x=mpg, fill=factor(cyl))) + geom_density(alpha=.7)
multiplot(p1, p2, cols = 2)
multiplot(p1, p2, cols = 2, panelnames = LETTERS)
multiplot(p1, p2, cols = 2, panelnames = LETTERS, panelnames.size = 24)
multiplot(p1, p2, p1, p2, layout=matrix(c(1,1,2,2,3,4), nc=3, byrow=FALSE), panelnames = LETTERS, panelnames.size = 24)
}

