% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genomeplot}
\alias{genomeplot}
\title{plot a copy number profile}
\usage{
genomeplot(segments, chrsizes)
}
\arguments{
\item{segments}{Data frame with segments from qdnaseq. Must include column chr, start, end and segmented, and *prob.}

\item{chrsizes}{Data frame with chromosome sizes.}
}
\value{
A ggplot object
}
\description{
plot a copy number profile
}
\examples{
# genomeplot(segments, chrsizes)
}

