% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCNArmSummary.R
\name{getCNArmSummary}
\alias{getCNArmSummary}
\title{Make a table of CN summary stats per chromosome arm}
\usage{
getCNArmSummary(segments, bandsFile)
}
\arguments{
\item{segments}{Data table with segments from qdnaseq.}

\item{bandsFile}{Path to cytobands.txt (see repo dakl/genome at github)}
}
\value{
A summary table per chromosome arm
}
\description{
Make a table of CN summary stats per chromosome arm
}
\examples{
#getCNArmSummary(segments, bandsFile = "cytobands.txt")
}

