% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGeneBodyCoverage.R
\name{getGeneBodyCoverage}
\alias{getGeneBodyCoverage}
\title{Get normalized gene body coverage for reports}
\usage{
getGeneBodyCoverage(reports, src = "RNASEQ_RNASEQ_METRICS")
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A data frame with normalized gene body coverage for all samples
}
\description{
Get normalized gene body coverage for reports
}
\examples{
# dat <- getGeneBodyCoverage(reports)
# ggplot(dat, aes(x=NORM_POS, y=NORM_COV, colour=REPORTID)) + geom_line()
}

