% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHSMetrics}
\alias{getHSMetrics}
\title{Get HS Metrics}
\usage{
getHSMetrics(reports, src = "PANEL_TUMOR_HSMETRICS")
}
\arguments{
\item{reports}{data frame with reports}

\item{src}{string with name of column name to use (from reports)}
}
\value{
A data table with hs metrics
}
\description{
Get HS Metrics
}
\examples{
#dat <- getHSMetrics(reports, src="PANEL_TUMOR_HSMETRICS")
#dat <- getHSMetrics(reports, src="PANEL_NORMAL_HSMETRICS")
#dat <- getHSMetrics(reports, src="RNASEQCAP_HSMETRICS")
}

