% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMarkDupsMetrics}
\alias{getMarkDupsMetrics}
\title{Get MarkDups Metrics}
\usage{
getMarkDupsMetrics(reports, src = "RNASeqCapMarkDupMetrics")
}
\arguments{
\item{reports}{data frame with reports}

\item{src}{string with name of column name to use (from reports)}
}
\value{
A data table with metrics from markDuplicates
}
\description{
Get MarkDups Metrics
}
\examples{
#dat <- getHSMetrics(reports, src="RNASeqCapHsMetrics")
#dat <- getHSMetrics(reports, src="TumorHsMetrics")
#dat <- getHSMetrics(reports, src="NormalHsMetrics")
}

