% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOncofuse.R
\name{getOncofuse}
\alias{getOncofuse}
\title{Get Oncofuse results}
\usage{
getOncofuse(reports, src = "RNASEQ_ONCOFUSE")
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A data table with Oncofuse results
}
\description{
Get Oncofuse results
}
\examples{
#dat <- getOncofuse(reports, src="RNASEQ_ONCOFUSE")
}

