% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getReports.R
\name{getReports}
\alias{getReports}
\title{Get a data frame with all reports found in path}
\usage{
getReports(path, recursive = TRUE, absolutepaths = FALSE, max)
}
\arguments{
\item{path}{Path to directory with reports.}

\item{recursive}{Boolean whether or not to recurse when looking for report files (default TRUE)}

\item{absolutepaths}{Boolean whether or not to make paths absolute paths instead of relative (default FALSE)}
}
\value{
A data frame with reports
}
\description{
Get a data frame with all reports found in path
}
\examples{
reports1 <- getReports(path="/proj/b2010040/private/clinseq/LAMLv6/", recursive=TRUE, absolutepaths=TRUE, max=10)
reports2 <- getReports(path="/proj/b2010040/private/clinseq/bla/", recursive=TRUE, absolutepaths=TRUE, max=10)
}

