% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSailfishMetrics}
\alias{getSailfishMetrics}
\title{Get metrics from sailfish}
\usage{
getSailfishMetrics(reports)
}
\arguments{
\item{reports}{data frame with reports}
}
\value{
A data frame with metrics from sailfish
}
\description{
Get metrics from sailfish
}
\examples{
#dat <- getSailfishMetrics(reports)
}

